Level cap 999 + XP/HP rework, boss bars, queen asset, mobile drag/freeze/buttons, click-swap
- Level cap 99->999. Levels 1-99 unchanged (existing XP/level preserved); 100+
  continue on a flattened XP ramp and linear HP so 999 is reachable and finite.
  Widened level to uint32 across camera field / AccountDB / progress.
- Unique mobs: up to 100 looters.
- Boss bars (new): top-centre HP bars for Super/Unique mobs in vision, name above
  + rarity below, unique first then by appearance, max 3.
- Queen Ant art now scales to its hitbox (was a fixed size).
- Mobile: freeze movement + disable joystick while a panel is open (fixes the
  stuck inventory drag); A/B/G moved to a bottom-right column that shifts up when
  the minimap expands.
- Loadout: short click on a slot swaps its main<->secondary pair (drag-tracked
  so it doesn't fire on drags).

+++ Server/Account/Database.hh
@@ -27,8 +27,8 @@ namespace AccountDB {
     int read_inventory(std::string const &user, std::vector<PetalStack> &out);
     int write_inventory(std::string const &user, std::vector<PetalStack> const &in);
 
-    int read_progress(std::string const &user, uint8_t &level, uint32_t &xp);
-    int write_progress(std::string const &user, uint8_t level, uint32_t xp);
+    int read_progress(std::string const &user, uint32_t &level, uint32_t &xp);
+    int write_progress(std::string const &user, uint32_t level, uint32_t xp);
 
     // Per-account mob kill tally (mob id + rarity -> count). Only killed combos
     // are stored. Callers batch a save() after crediting a death.
